Allow BitPaddedInt to return long values
Ok, this is not what I'd call a nice change, but I have a file with POPM frame where the play count doesn't fit the int type on 32-bit systems, so the value needs to be stored as long. The value is obviously wrong, but since the file is technically valid (ID3 allows it to be even larger than 32-bit unsigned int), I think mutagen should be able handle it.